requirements.txt
Flask
Pillow
rembg
gunicorn
